Update test references
