+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTAnyTypeDeclaration.java
@@ -36,7 +36,8 @@ public interface ASTAnyTypeDeclaration
             ASTBodyDeclaration,
             ASTTopLevelDeclaration,
             FinalizableNode,
-            JavadocCommentOwner {
+            JavadocCommentOwner,
+            NonLocalDeclarationNode {
 
     @Override
     @NonNull
