+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTFieldDeclaration.java
@@ -28,7 +28,8 @@ public final class ASTFieldDeclaration extends AbstractJavaNode
                AccessNode,
                ASTBodyDeclaration,
                InternalInterfaces.MultiVariableIdOwner,
-               JavadocCommentOwner {
+               JavadocCommentOwner,
+               NonLocalDeclarationNode {
 
 
     ASTFieldDeclaration(int id) {
