+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTMethodOrConstructorDeclaration.java
@@ -31,7 +31,8 @@ public interface ASTMethodOrConstructorDeclaration
             ASTBodyDeclaration,
             TypeParamOwnerNode,
             GenericNode<JavaNode>,
-            JavadocCommentOwner {
+            JavadocCommentOwner,
+            NonLocalDeclarationNode {
 
 
     @Override
