+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/Bug1429.txt
@@ -1,8 +1,8 @@
 +- CompilationUnit[@PackageName = ""]
-   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "Bug1429", @CanonicalName = "Bug1429", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "Bug1429", @TopLevel = "true", @Visibility = "public"]
+   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "Bug1429", @CanonicalName = "Bug1429", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "Bug1429", @Static = "false", @TopLevel = "true", @Visibility = "public"]
       +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
       +- ClassOrInterfaceBody[@Empty = "false", @Size = "1"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "getAttributeTuples", @Name = "getAttributeTuples", @Overridden = "false", @Varargs = "false", @Visibility = "public", @Void = "false"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "getAttributeTuples", @Name = "getAttributeTuples", @Overridden = "false", @Static = "false", @Varargs = "false", @Visibility = "public", @Void = "false"]
             +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
             +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "Set"]
             |  +- TypeArguments[@Diamond = "false", @Empty = "false", @Size = "1"]
@@ -39,10 +39,10 @@
                                     +- ConstructorCall[@AnonymousClass = "true", @CompileTimeConstant = "false", @DiamondTypeArgs = "false", @MethodName = "new", @ParenthesisDepth = "0", @Parenthesized = "false", @QualifiedInstanceCreation = "false"]
                                        +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "Transformer"]
                                        +- ArgumentList[@Empty = "true", @Size = "0"]
-                                       +- AnonymousClassDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "true", @BinaryName = "Bug1429$1", @CanonicalName = null, @EffectiveVisibility = "anonymous", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @Record = "false", @RegularClass = "false", @RegularInterface = "false", @SimpleName = "", @TopLevel = "false", @Visibility = "anonymous"]
+                                       +- AnonymousClassDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "true", @BinaryName = "Bug1429$1", @CanonicalName = null, @EffectiveVisibility = "anonymous", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @Record = "false", @RegularClass = "false", @RegularInterface = "false", @SimpleName = "", @Static = "false", @TopLevel = "false", @Visibility = "anonymous"]
                                           +- ModifierList[@EffectiveModifiers = "{}", @ExplicitModifiers = "{}"]
                                           +- ClassOrInterfaceBody[@Empty = "false", @Size = "1"]
-                                             +- MethodDeclaration[@Abstract = "false", @Arity = "1", @EffectiveVisibility = "anonymous", @Image = "transform", @Name = "transform", @Overridden = "true", @Varargs = "false", @Visibility = "public", @Void = "false"]
+                                             +- MethodDeclaration[@Abstract = "false", @Arity = "1", @EffectiveVisibility = "anonymous", @Image = "transform", @Name = "transform", @Overridden = "true", @Static = "false", @Varargs = "false", @Visibility = "public", @Void = "false"]
                                                 +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
                                                 |  +- Annotation[@SimpleName = "Override"]
                                                 |     +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "Override"]
