+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/Bug1530.txt
@@ -1,14 +1,14 @@
 +- CompilationUnit[@PackageName = ""]
-   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "Bug1530", @CanonicalName = "Bug1530", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "Bug1530", @TopLevel = "true", @Visibility = "public"]
+   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "Bug1530", @CanonicalName = "Bug1530", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "Bug1530", @Static = "false", @TopLevel = "true", @Visibility = "public"]
       +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
       +- ClassOrInterfaceBody[@Empty = "false", @Size = "1"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "incChild", @Name = "incChild", @Overridden = "false", @Varargs = "false", @Visibility = "public", @Void = "true"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "incChild", @Name = "incChild", @Overridden = "false", @Static = "false", @Varargs = "false", @Visibility = "public", @Void = "true"]
             +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
             +- VoidType[]
             +- FormalParameters[@Empty = "true", @Size = "0"]
             +- Block[@Empty = "false", @Size = "1", @containsComment = "false"]
                +- ExpressionStatement[]
-                  +- UnaryExpression[@CompileTimeConstant = "false", @Operator = "++", @ParenthesisDepth = "0", @Parenthesized = "false"]
+                  +- UnaryExpression[@CompileTimeConstant = "false", @Operator = "++", @ParenthesisDepth = "0", @Parenthesized = "false", @Prefix = "false"]
                      +- FieldAccess[@AccessType = "WRITE", @CompileTimeConstant = "false", @Image = "currentChild", @Name = "currentChild", @ParenthesisDepth = "0", @Parenthesized = "false"]
                         +- CastExpression[@CompileTimeConstant = "false", @ParenthesisDepth = "1", @Parenthesized = "true"]
                            +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "PathElement"]
