+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/EmptyStmts3.txt
@@ -2,7 +2,7 @@
    +- PackageDeclaration[@Name = "c"]
    |  +- ModifierList[@EffectiveModifiers = "{}", @ExplicitModifiers = "{}"]
    +- ImportDeclaration[@ImportOnDemand = "false", @ImportedName = "a", @ImportedSimpleName = "a", @PackageName = "", @Static = "false"]
-   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "c.Foo", @CanonicalName = "c.Foo", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "c", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "Foo", @TopLevel = "true", @Visibility = "public"]
+   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "c.Foo", @CanonicalName = "c.Foo", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "c", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "Foo", @Static = "false", @TopLevel = "true", @Visibility = "public"]
    |  +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
    |  +- ClassOrInterfaceBody[@Empty = "true", @Size = "0"]
    +- EmptyDeclaration[]
