+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/GitHubBug207.txt
@@ -1,8 +1,8 @@
 +- CompilationUnit[@PackageName = ""]
-   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "GitHubBug207", @CanonicalName = "GitHubBug207", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "GitHubBug207", @TopLevel = "true", @Visibility = "public"]
+   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "GitHubBug207", @CanonicalName = "GitHubBug207", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "GitHubBug207", @Static = "false", @TopLevel = "true", @Visibility = "public"]
       +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
       +- ClassOrInterfaceBody[@Empty = "false", @Size = "1"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "1", @EffectiveVisibility = "private", @Image = "resourceHttpMessageWriter", @Name = "resourceHttpMessageWriter", @Overridden = "false", @Varargs = "false", @Visibility = "private", @Void = "false"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "1", @EffectiveVisibility = "private", @Image = "resourceHttpMessageWriter", @Name = "resourceHttpMessageWriter", @Overridden = "false", @Static = "true", @Varargs = "false", @Visibility = "private", @Void = "false"]
             +- ModifierList[@EffectiveModifiers = "{private, static}", @ExplicitModifiers = "{private, static}"]
             +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "HttpMessageWriter"]
             |  +- TypeArguments[@Diamond = "false", @Empty = "false", @Size = "1"]
