+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/GitHubBug208.txt
@@ -1,21 +1,21 @@
 +- CompilationUnit[@PackageName = ""]
-   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "GitHubBug208", @CanonicalName = "GitHubBug208", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "GitHubBug208", @TopLevel = "true", @Visibility = "public"]
+   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "GitHubBug208", @CanonicalName = "GitHubBug208", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "GitHubBug208", @Static = "false", @TopLevel = "true", @Visibility = "public"]
       +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
       +- ClassOrInterfaceBody[@Empty = "false", @Size = "1"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "testMethod", @Name = "testMethod", @Overridden = "false", @Varargs = "false", @Visibility = "public", @Void = "true"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "testMethod", @Name = "testMethod", @Overridden = "false", @Static = "false", @Varargs = "false", @Visibility = "public", @Void = "true"]
             +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
             +- VoidType[]
             +- FormalParameters[@Empty = "true", @Size = "0"]
             +- Block[@Empty = "false", @Size = "1", @containsComment = "false"]
                +- LocalClassStatement[]
-                  +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "GitHubBug208$1LocalClass", @CanonicalName = null, @EffectiveVisibility = "local", @Enum = "false", @Final = "false", @Interface = "false", @Local = "true", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "LocalClass", @TopLevel = "false", @Visibility = "local"]
+                  +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "GitHubBug208$1LocalClass", @CanonicalName = null, @EffectiveVisibility = "local", @Enum = "false", @Final = "false", @Interface = "false", @Local = "true", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "LocalClass", @Static = "false", @TopLevel = "false", @Visibility = "local"]
                      +- ModifierList[@EffectiveModifiers = "{}", @ExplicitModifiers = "{}"]
                      |  +- Annotation[@SimpleName = "Lazy"]
                      |  |  +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "Lazy"]
                      |  +- Annotation[@SimpleName = "Configuration"]
                      |     +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "Configuration"]
                      +- ClassOrInterfaceBody[@Empty = "false", @Size = "1"]
-                        +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "local", @Image = "foo", @Name = "foo", @Overridden = "false", @Varargs = "false", @Visibility = "package", @Void = "false"]
+                        +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "local", @Image = "foo", @Name = "foo", @Overridden = "false", @Static = "false", @Varargs = "false", @Visibility = "package", @Void = "false"]
                            +- ModifierList[@EffectiveModifiers = "{}", @ExplicitModifiers = "{}"]
                            |  +- Annotation[@SimpleName = "Bean"]
                            |     +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "Bean"]
