+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/LambdaBug1333.txt
@@ -1,8 +1,8 @@
 +- CompilationUnit[@PackageName = ""]
-   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "Bug1333", @CanonicalName = "Bug1333", @EffectiveVisibility = "package", @Enum = "false", @Final = "true", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "true", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "Bug1333", @TopLevel = "true", @Visibility = "package"]
+   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "Bug1333", @CanonicalName = "Bug1333", @EffectiveVisibility = "package", @Enum = "false", @Final = "true", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "true", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "Bug1333", @Static = "false", @TopLevel = "true", @Visibility = "package"]
       +- ModifierList[@EffectiveModifiers = "{final}", @ExplicitModifiers = "{final}"]
       +- ClassOrInterfaceBody[@Empty = "false", @Size = "4"]
-         +- FieldDeclaration[@EffectiveVisibility = "private", @Visibility = "private"]
+         +- FieldDeclaration[@EffectiveVisibility = "private", @Static = "true", @Visibility = "private"]
          |  +- ModifierList[@EffectiveModifiers = "{private, static, final}", @ExplicitModifiers = "{private, static, final}"]
          |  +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "Logger"]
          |  +- VariableDeclarator[@Initializer = "true", @Name = "LOG"]
@@ -12,7 +12,7 @@
          |        +- ArgumentList[@Empty = "false", @Size = "1"]
          |           +- ClassLiteral[@CompileTimeConstant = "false", @ParenthesisDepth = "0", @Parenthesized = "false"]
          |              +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "Foo"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "package", @Image = "deleteDirectoriesByNamePattern", @Name = "deleteDirectoriesByNamePattern", @Overridden = "false", @Varargs = "false", @Visibility = "public", @Void = "true"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "package", @Image = "deleteDirectoriesByNamePattern", @Name = "deleteDirectoriesByNamePattern", @Overridden = "false", @Static = "false", @Varargs = "false", @Visibility = "public", @Void = "true"]
          |  +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
          |  +- VoidType[]
          |  +- FormalParameters[@Empty = "true", @Size = "0"]
@@ -28,7 +28,7 @@
          |                 +- MethodCall[@CompileTimeConstant = "false", @Image = "deleteDirectory", @MethodName = "deleteDirectory", @ParenthesisDepth = "0", @Parenthesized = "false"]
          |                    +- ArgumentList[@Empty = "false", @Size = "1"]
          |                       +- VariableAccess[@AccessType = "READ", @CompileTimeConstant = "false", @Image = "path", @Name = "path", @ParenthesisDepth = "0", @Parenthesized = "false"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "1", @EffectiveVisibility = "private", @Image = "delete", @Name = "delete", @Overridden = "false", @Varargs = "false", @Visibility = "private", @Void = "true"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "1", @EffectiveVisibility = "private", @Image = "delete", @Name = "delete", @Overridden = "false", @Static = "false", @Varargs = "false", @Visibility = "private", @Void = "true"]
          |  +- ModifierList[@EffectiveModifiers = "{private}", @ExplicitModifiers = "{private}"]
          |  +- VoidType[]
          |  +- FormalParameters[@Empty = "false", @Size = "1"]
@@ -47,7 +47,7 @@
          |              +- MethodCall[@CompileTimeConstant = "false", @Image = "toString", @MethodName = "toString", @ParenthesisDepth = "0", @Parenthesized = "false"]
          |                 +- VariableAccess[@AccessType = "READ", @CompileTimeConstant = "false", @Image = "consumer", @Name = "consumer", @ParenthesisDepth = "0", @Parenthesized = "false"]
          |                 +- ArgumentList[@Empty = "true", @Size = "0"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "1", @EffectiveVisibility = "private", @Image = "deleteDirectory", @Name = "deleteDirectory", @Overridden = "false", @Varargs = "false", @Visibility = "private", @Void = "true"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "1", @EffectiveVisibility = "private", @Image = "deleteDirectory", @Name = "deleteDirectory", @Overridden = "false", @Static = "false", @Varargs = "false", @Visibility = "private", @Void = "true"]
             +- ModifierList[@EffectiveModifiers = "{private}", @ExplicitModifiers = "{private}"]
             +- VoidType[]
             +- FormalParameters[@Empty = "false", @Size = "1"]
