+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/LambdaBug1470.txt
@@ -3,15 +3,15 @@
    |  +- ModifierList[@EffectiveModifiers = "{}", @ExplicitModifiers = "{}"]
    +- ImportDeclaration[@ImportOnDemand = "false", @ImportedName = "rx.Observable", @ImportedSimpleName = "Observable", @PackageName = "rx", @Static = "false"]
    +- ImportDeclaration[@ImportOnDemand = "false", @ImportedName = "rx.Subscriber", @ImportedSimpleName = "Subscriber", @PackageName = "rx", @Static = "false"]
-   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "com.sample.test.pmdTest", @CanonicalName = "com.sample.test.pmdTest", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "com.sample.test", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "pmdTest", @TopLevel = "true", @Visibility = "public"]
+   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "com.sample.test.pmdTest", @CanonicalName = "com.sample.test.pmdTest", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "com.sample.test", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "pmdTest", @Static = "false", @TopLevel = "true", @Visibility = "public"]
       +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
       +- ClassOrInterfaceBody[@Empty = "false", @Size = "3"]
-         +- FieldDeclaration[@EffectiveVisibility = "private", @Visibility = "private"]
+         +- FieldDeclaration[@EffectiveVisibility = "private", @Static = "false", @Visibility = "private"]
          |  +- ModifierList[@EffectiveModifiers = "{private}", @ExplicitModifiers = "{private}"]
          |  +- PrimitiveType[@Kind = "boolean"]
          |  +- VariableDeclarator[@Initializer = "false", @Name = "stuff"]
          |     +- VariableDeclaratorId[@ArrayType = "false", @EffectiveVisibility = "private", @EnumConstant = "false", @ExceptionBlockParameter = "false", @Field = "true", @Final = "false", @ForLoopVariable = "false", @ForeachVariable = "false", @FormalParameter = "false", @LambdaParameter = "false", @LocalVariable = "false", @Name = "stuff", @PatternBinding = "false", @RecordComponent = "false", @ResourceDeclaration = "false", @TypeInferred = "false", @Visibility = "private"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "testSuper", @Name = "testSuper", @Overridden = "false", @Varargs = "false", @Visibility = "public", @Void = "false"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "testSuper", @Name = "testSuper", @Overridden = "false", @Static = "false", @Varargs = "false", @Visibility = "public", @Void = "false"]
          |  +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
          |  +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "Observable"]
          |  |  +- TypeArguments[@Diamond = "false", @Empty = "false", @Size = "1"]
@@ -45,7 +45,7 @@
          |                 |     +- ModifierList[@EffectiveModifiers = "{}", @ExplicitModifiers = "{}"]
          |                 |     +- VariableDeclaratorId[@ArrayType = "false", @EffectiveVisibility = "package", @EnumConstant = "false", @ExceptionBlockParameter = "false", @Field = "false", @Final = "false", @ForLoopVariable = "false", @ForeachVariable = "false", @FormalParameter = "true", @LambdaParameter = "true", @LocalVariable = "false", @Name = "authToken", @PatternBinding = "false", @RecordComponent = "false", @ResourceDeclaration = "false", @TypeInferred = "true", @Visibility = "package"]
          |                 +- BooleanLiteral[@CompileTimeConstant = "true", @ParenthesisDepth = "0", @Parenthesized = "false", @True = "false"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "testSuper2", @Name = "testSuper2", @Overridden = "false", @Varargs = "false", @Visibility = "public", @Void = "false"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "testSuper2", @Name = "testSuper2", @Overridden = "false", @Static = "false", @Varargs = "false", @Visibility = "public", @Void = "false"]
             +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
             +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "Observable"]
             |  +- TypeArguments[@Diamond = "false", @Empty = "false", @Size = "1"]
