+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/LambdaBug206.txt
@@ -1,8 +1,8 @@
 +- CompilationUnit[@PackageName = ""]
-   +- AnnotationTypeDeclaration[@Abstract = "true", @Annotation = "true", @Anonymous = "false", @BinaryName = "Foo", @CanonicalName = "Foo", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "true", @Local = "false", @Nested = "false", @PackageName = "", @Record = "false", @RegularClass = "false", @RegularInterface = "false", @SimpleName = "Foo", @TopLevel = "true", @Visibility = "public"]
+   +- AnnotationTypeDeclaration[@Abstract = "true", @Annotation = "true", @Anonymous = "false", @BinaryName = "Foo", @CanonicalName = "Foo", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "true", @Local = "false", @Nested = "false", @PackageName = "", @Record = "false", @RegularClass = "false", @RegularInterface = "false", @SimpleName = "Foo", @Static = "false", @TopLevel = "true", @Visibility = "public"]
       +- ModifierList[@EffectiveModifiers = "{public, abstract}", @ExplicitModifiers = "{public}"]
       +- AnnotationTypeBody[@Empty = "false", @Size = "1"]
-         +- FieldDeclaration[@EffectiveVisibility = "public", @Visibility = "public"]
+         +- FieldDeclaration[@EffectiveVisibility = "public", @Static = "true", @Visibility = "public"]
             +- ModifierList[@EffectiveModifiers = "{public, static, final}", @ExplicitModifiers = "{static, final}"]
             +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "ThreadLocal"]
             |  +- TypeArguments[@Diamond = "false", @Empty = "false", @Size = "1"]
