+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/SwitchStatements.txt
@@ -1,8 +1,8 @@
 +- CompilationUnit[@PackageName = ""]
-   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "SwitchStatements", @CanonicalName = "SwitchStatements", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "SwitchStatements", @TopLevel = "true", @Visibility = "public"]
+   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "SwitchStatements", @CanonicalName = "SwitchStatements", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "SwitchStatements", @Static = "false", @TopLevel = "true", @Visibility = "public"]
       +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
       +- ClassOrInterfaceBody[@Empty = "false", @Size = "1"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "myMethod", @Name = "myMethod", @Overridden = "false", @Varargs = "false", @Visibility = "public", @Void = "true"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Image = "myMethod", @Name = "myMethod", @Overridden = "false", @Static = "false", @Varargs = "false", @Visibility = "public", @Void = "true"]
             +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
             +- VoidType[]
             +- FormalParameters[@Empty = "true", @Size = "0"]
