+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/SynchronizedStmts.txt
@@ -1,8 +1,8 @@
 +- CompilationUnit[@PackageName = ""]
-   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "Sync", @CanonicalName = "Sync", @EffectiveVisibility = "package", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "true", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "Sync", @TopLevel = "true", @Visibility = "package"]
+   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "Sync", @CanonicalName = "Sync", @EffectiveVisibility = "package", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "true", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "Sync", @Static = "false", @TopLevel = "true", @Visibility = "package"]
       +- ModifierList[@EffectiveModifiers = "{}", @ExplicitModifiers = "{}"]
       +- ClassOrInterfaceBody[@Empty = "false", @Size = "1"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "package", @Image = "getInstance", @Name = "getInstance", @Overridden = "false", @Varargs = "false", @Visibility = "public", @Void = "true"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "package", @Image = "getInstance", @Name = "getInstance", @Overridden = "false", @Static = "true", @Varargs = "false", @Visibility = "public", @Void = "true"]
             +- ModifierList[@EffectiveModifiers = "{public, static}", @ExplicitModifiers = "{public, static}"]
             +- VoidType[]
             +- FormalParameters[@Empty = "true", @Size = "0"]
