+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/jdkversiontests/java14/YieldStatements.txt
@@ -1,5 +1,5 @@
 +- CompilationUnit[@PackageName = ""]
-   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "YieldStatements", @CanonicalName = "YieldStatements", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "YieldStatements", @TopLevel = "true", @Visibility = "public"]
+   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "YieldStatements", @CanonicalName = "YieldStatements", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "YieldStatements", @Static = "false", @TopLevel = "true", @Visibility = "public"]
       +- ModifierList[@EffectiveModifiers = "{public}", @ExplicitModifiers = "{public}"]
       +- ClassOrInterfaceBody[@Empty = "false", @Size = "1"]
          +- Initializer[@Static = "false"]
@@ -44,16 +44,16 @@
                               +- YieldStatement[]
                               |  +- NumericLiteral[@Base = "10", @CompileTimeConstant = "true", @DoubleLiteral = "false", @FloatLiteral = "false", @Image = "2", @IntLiteral = "true", @Integral = "true", @LongLiteral = "false", @ParenthesisDepth = "1", @Parenthesized = "true", @ValueAsDouble = "2.0", @ValueAsFloat = "2.0", @ValueAsInt = "2", @ValueAsLong = "2"]
                               +- YieldStatement[]
-                              |  +- UnaryExpression[@CompileTimeConstant = "false", @Operator = "++", @ParenthesisDepth = "0", @Parenthesized = "false"]
+                              |  +- UnaryExpression[@CompileTimeConstant = "false", @Operator = "++", @ParenthesisDepth = "0", @Parenthesized = "false", @Prefix = "true"]
                               |     +- VariableAccess[@AccessType = "WRITE", @CompileTimeConstant = "false", @Image = "bar", @Name = "bar", @ParenthesisDepth = "0", @Parenthesized = "false"]
                               +- YieldStatement[]
-                              |  +- UnaryExpression[@CompileTimeConstant = "false", @Operator = "--", @ParenthesisDepth = "0", @Parenthesized = "false"]
+                              |  +- UnaryExpression[@CompileTimeConstant = "false", @Operator = "--", @ParenthesisDepth = "0", @Parenthesized = "false", @Prefix = "true"]
                               |     +- VariableAccess[@AccessType = "WRITE", @CompileTimeConstant = "false", @Image = "bar", @Name = "bar", @ParenthesisDepth = "0", @Parenthesized = "false"]
                               +- ExpressionStatement[]
-                              |  +- UnaryExpression[@CompileTimeConstant = "false", @Operator = "++", @ParenthesisDepth = "0", @Parenthesized = "false"]
+                              |  +- UnaryExpression[@CompileTimeConstant = "false", @Operator = "++", @ParenthesisDepth = "0", @Parenthesized = "false", @Prefix = "false"]
                               |     +- VariableAccess[@AccessType = "WRITE", @CompileTimeConstant = "false", @Image = "yield", @Name = "yield", @ParenthesisDepth = "0", @Parenthesized = "false"]
                               +- ExpressionStatement[]
-                              |  +- UnaryExpression[@CompileTimeConstant = "false", @Operator = "--", @ParenthesisDepth = "0", @Parenthesized = "false"]
+                              |  +- UnaryExpression[@CompileTimeConstant = "false", @Operator = "--", @ParenthesisDepth = "0", @Parenthesized = "false", @Prefix = "false"]
                               |     +- VariableAccess[@AccessType = "WRITE", @CompileTimeConstant = "false", @Image = "yield", @Name = "yield", @ParenthesisDepth = "0", @Parenthesized = "false"]
                               +- IfStatement[@Else = "true"]
                               |  +- BooleanLiteral[@CompileTimeConstant = "true", @ParenthesisDepth = "0", @Parenthesized = "false", @True = "true"]
