+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/jdkversiontests/java16/Point.txt
@@ -1,5 +1,5 @@
 +- CompilationUnit[@PackageName = ""]
-   +- RecordDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "Point", @CanonicalName = "Point", @EffectiveVisibility = "public", @Enum = "false", @Final = "true", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @Record = "true", @RegularClass = "false", @RegularInterface = "false", @SimpleName = "Point", @TopLevel = "true", @Visibility = "public"]
+   +- RecordDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "Point", @CanonicalName = "Point", @EffectiveVisibility = "public", @Enum = "false", @Final = "true", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @Record = "true", @RegularClass = "false", @RegularInterface = "false", @SimpleName = "Point", @Static = "false", @TopLevel = "true", @Visibility = "public"]
       +- ModifierList[@EffectiveModifiers = "{public, final}", @ExplicitModifiers = "{public}"]
       +- RecordComponentList[@Empty = "false", @Size = "2", @Varargs = "false"]
       |  +- RecordComponent[@EffectiveVisibility = "private", @Varargs = "false", @Visibility = "private"]
@@ -11,7 +11,7 @@
       |     +- PrimitiveType[@Kind = "int"]
       |     +- VariableDeclaratorId[@ArrayType = "false", @EffectiveVisibility = "private", @EnumConstant = "false", @ExceptionBlockParameter = "false", @Field = "false", @Final = "true", @ForLoopVariable = "false", @ForeachVariable = "false", @FormalParameter = "false", @LambdaParameter = "false", @LocalVariable = "false", @Name = "y", @PatternBinding = "false", @RecordComponent = "true", @ResourceDeclaration = "false", @TypeInferred = "false", @Visibility = "private"]
       +- RecordBody[@Empty = "false", @Size = "1"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "1", @EffectiveVisibility = "public", @Image = "main", @MainMethod = "true", @Name = "main", @Overridden = "false", @Varargs = "false", @Visibility = "public", @Void = "true"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "1", @EffectiveVisibility = "public", @Image = "main", @MainMethod = "true", @Name = "main", @Overridden = "false", @Static = "true", @Varargs = "false", @Visibility = "public", @Void = "true"]
             +- ModifierList[@EffectiveModifiers = "{public, static}", @ExplicitModifiers = "{public, static}"]
             +- VoidType[]
             +- FormalParameters[@Empty = "false", @Size = "1"]
