+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/jdkversiontests/java17/LocalVars.txt
@@ -1,18 +1,18 @@
-+- CompilationUnit[@PackageName = "", @declarationsAreInDefaultPackage = "true"]
-   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "LocalVars", @CanonicalName = "LocalVars", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Image = "LocalVars", @Interface = "false", @Local = "false", @Native = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Private = "false", @Protected = "false", @Public = "true", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "LocalVars", @Static = "false", @Strictfp = "false", @Synchronized = "false", @SyntacticallyAbstract = "false", @SyntacticallyFinal = "false", @SyntacticallyPublic = "true", @SyntacticallyStatic = "false", @TopLevel = "true", @Transient = "false", @Visibility = "public", @Volatile = "false"]
++- CompilationUnit[@PackageName = ""]
+   +- ClassOrInterfaceDeclaration[@Abstract = "false", @Annotation = "false", @Anonymous = "false", @BinaryName = "LocalVars", @CanonicalName = "LocalVars", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Image = "LocalVars", @Interface = "false", @Local = "false", @Nested = "false", @PackageName = "", @PackagePrivate = "false", @Record = "false", @RegularClass = "true", @RegularInterface = "false", @SimpleName = "LocalVars", @Static = "false", @TopLevel = "true", @Visibility = "public"]
       +- ModifierList[]
       +- ClassOrInterfaceBody[@Empty = "false", @Size = "1"]
-         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Final = "false", @Image = "aMethod", @MainMethod = "false", @MethodName = "aMethod", @Name = "aMethod", @Native = "false", @Overridden = "false", @PackagePrivate = "false", @Private = "false", @Protected = "false", @Public = "true", @Static = "false", @Strictfp = "false", @Synchronized = "false", @SyntacticallyAbstract = "false", @SyntacticallyFinal = "false", @SyntacticallyPublic = "true", @SyntacticallyStatic = "false", @Transient = "false", @Varargs = "false", @Visibility = "public", @Void = "true", @Volatile = "false"]
+         +- MethodDeclaration[@Abstract = "false", @Arity = "0", @EffectiveVisibility = "public", @Final = "false", @Image = "aMethod", @MainMethod = "false", @Name = "aMethod", @Overridden = "false", @Static = "false", @Varargs = "false", @Visibility = "public", @Void = "true"]
             +- ModifierList[]
-            +- VoidType[@ArrayDepth = "0", @ArrayType = "false", @ClassOrInterfaceType = "false", @PrimitiveType = "false", @TypeImage = "void"]
+            +- VoidType[]
             +- FormalParameters[@Empty = "true", @Size = "0"]
             +- Block[@Empty = "false", @Size = "2", @containsComment = "true"]
-               +- LocalVariableDeclaration[@Abstract = "false", @EffectiveVisibility = "local", @Final = "false", @Native = "false", @PackagePrivate = "false", @Private = "false", @Protected = "false", @Public = "false", @Static = "false", @Strictfp = "false", @Synchronized = "false", @SyntacticallyAbstract = "false", @SyntacticallyFinal = "false", @SyntacticallyPublic = "false", @SyntacticallyStatic = "false", @Transient = "false", @TypeInferred = "false", @Visibility = "local", @Volatile = "false"]
+               +- LocalVariableDeclaration[@EffectiveVisibility = "local", @Final = "false", @TypeInferred = "false", @Visibility = "local"]
                |  +- ModifierList[]
-               |  +- ClassOrInterfaceType[@ArrayDepth = "0", @ArrayType = "false", @ClassOrInterfaceType = "true", @FullyQualified = "false", @PrimitiveType = "false", @SimpleName = "String", @TypeImage = "String"]
+               |  +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "String"]
                |  +- VariableDeclarator[@Initializer = "true", @Name = "sealed"]
-               |     +- VariableDeclaratorId[@Abstract = "false", @ArrayType = "false", @EffectiveVisibility = "local", @EnumConstant = "false", @ExceptionBlockParameter = "false", @Field = "false", @Final = "false", @ForLoopVariable = "false", @ForeachVariable = "false", @FormalParameter = "false", @Image = "sealed", @LambdaParameter = "false", @LocalVariable = "true", @Name = "sealed", @Native = "false", @PackagePrivate = "false", @PatternBinding = "false", @Private = "false", @Protected = "false", @Public = "false", @RecordComponent = "false", @ResourceDeclaration = "false", @Static = "false", @Strictfp = "false", @Synchronized = "false", @SyntacticallyAbstract = "false", @SyntacticallyFinal = "false", @SyntacticallyPublic = "false", @SyntacticallyStatic = "false", @Transient = "false", @TypeInferred = "false", @VariableName = "sealed", @Visibility = "local", @Volatile = "false"]
-               |     +- NullLiteral[@BooleanLiteral = "false", @CharLiteral = "false", @CompileTimeConstant = "false", @DoubleLiteral = "false", @Expression = "true", @FloatLiteral = "false", @IntLiteral = "false", @LongLiteral = "false", @NullLiteral = "true", @NumericLiteral = "false", @ParenthesisDepth = "0", @Parenthesized = "false", @StringLiteral = "false"]
+               |     +- VariableDeclaratorId[@ArrayType = "false", @EffectiveVisibility = "local", @EnumConstant = "false", @ExceptionBlockParameter = "false", @Field = "false", @Final = "false", @ForLoopVariable = "false", @ForeachVariable = "false", @FormalParameter = "false", @LambdaParameter = "false", @LocalVariable = "true", @Name = "sealed", @PatternBinding = "false", @RecordComponent = "false", @ResourceDeclaration = "false", @TypeInferred = "false", @Visibility = "local"]
+               |     +- NullLiteral[@CompileTimeConstant = "false", @Expression = "true", @ParenthesisDepth = "0", @Parenthesized = "false"]
                +- ExpressionStatement[]
                   +- AssignmentExpression[@CompileTimeConstant = "false", @Compound = "false", @Expression = "true", @Operator = "=", @ParenthesisDepth = "0", @Parenthesized = "false"]
                      +- VariableAccess[@AccessType = "WRITE", @CompileTimeConstant = "false", @Expression = "true", @Image = "sealed", @Name = "sealed", @ParenthesisDepth = "0", @Parenthesized = "false"]
