+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/ast/jdkversiontests/java17/expression/Expr.txt
@@ -1,11 +1,11 @@
-+- CompilationUnit[@PackageName = "com.example.expression", @declarationsAreInDefaultPackage = "false"]
++- CompilationUnit[@PackageName = "com.example.expression"]
    +- PackageDeclaration[@Name = "com.example.expression"]
    |  +- ModifierList[]
-   +- ClassOrInterfaceDeclaration[@Abstract = "true", @Annotation = "false", @Anonymous = "false", @BinaryName = "com.example.expression.Expr", @CanonicalName = "com.example.expression.Expr", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Image = "Expr", @Interface = "true", @Local = "false", @Native = "false", @Nested = "false", @PackageName = "com.example.expression", @PackagePrivate = "false", @Private = "false", @Protected = "false", @Public = "true", @Record = "false", @RegularClass = "false", @RegularInterface = "true", @SimpleName = "Expr", @Static = "false", @Strictfp = "false", @Synchronized = "false", @SyntacticallyAbstract = "false", @SyntacticallyFinal = "false", @SyntacticallyPublic = "true", @SyntacticallyStatic = "false", @TopLevel = "true", @Transient = "false", @Visibility = "public", @Volatile = "false"]
+   +- ClassOrInterfaceDeclaration[@Abstract = "true", @Annotation = "false", @Anonymous = "false", @BinaryName = "com.example.expression.Expr", @CanonicalName = "com.example.expression.Expr", @EffectiveVisibility = "public", @Enum = "false", @Final = "false", @Image = "Expr", @Interface = "true", @Local = "false", @Nested = "false", @PackageName = "com.example.expression", @PackagePrivate = "false", @Record = "false", @RegularClass = "false", @RegularInterface = "true", @SimpleName = "Expr", @Static = "false", @TopLevel = "true", @Visibility = "public"]
       +- ModifierList[]
       +- PermitsList[@Empty = "false", @Size = "4"]
-      |  +- ClassOrInterfaceType[@ArrayDepth = "0", @ArrayType = "false", @ClassOrInterfaceType = "true", @FullyQualified = "false", @PrimitiveType = "false", @SimpleName = "ConstantExpr", @TypeImage = "ConstantExpr"]
-      |  +- ClassOrInterfaceType[@ArrayDepth = "0", @ArrayType = "false", @ClassOrInterfaceType = "true", @FullyQualified = "false", @PrimitiveType = "false", @SimpleName = "PlusExpr", @TypeImage = "PlusExpr"]
-      |  +- ClassOrInterfaceType[@ArrayDepth = "0", @ArrayType = "false", @ClassOrInterfaceType = "true", @FullyQualified = "false", @PrimitiveType = "false", @SimpleName = "TimesExpr", @TypeImage = "TimesExpr"]
-      |  +- ClassOrInterfaceType[@ArrayDepth = "0", @ArrayType = "false", @ClassOrInterfaceType = "true", @FullyQualified = "false", @PrimitiveType = "false", @SimpleName = "NegExpr", @TypeImage = "NegExpr"]
+      |  +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "ConstantExpr"]
+      |  +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "PlusExpr"]
+      |  +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "TimesExpr"]
+      |  +- ClassOrInterfaceType[@FullyQualified = "false", @SimpleName = "NegExpr"]
       +- ClassOrInterfaceBody[@Empty = "true", @Size = "0"]
